requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
notebook
mysql-connector-python
python-dotenv
xgboost
imblearn
ensure
pyYAML
joblib
types-PyYAML
python-box
streamlit
# -e .